Milestone 2 hardening: narrow Grid/RLE types for embedded memory footprint
Grid/Tile data and RLEbits shrink from int to uint8_t/uint16_t (Cell type),
cutting grid/tile memory 4x ahead of the eventual ESP32 target. Updates
main.cpp and test_main.cpp call sites to match, including a sizeof(int) ->
sizeof(Cell) fix that would have silently overstated the compression ratio.
Recorded in Discussion.md.

+++ include/rle.h
@@ -1,19 +1,25 @@
 #ifndef RLE_H
 #define RLE_H
 #include <vector>
+#include <cstdint>
+#include <cstddef>
 namespace compressor{
+	// A run of `count` consecutive cells of `value`.
+	// value is uint8_t (matches Grid/Tile's Cell type: 0=free, 1=obstacle, 2=uncertain).
+	// count is uint16_t, not uint8_t, because a single run can span an entire
+	// 100x100 grid (10000 cells) -- that doesn't fit in 8 bits but fits easily in 16.
 	struct RLEbits {
-		int value;
-		int count;
+		uint8_t value;
+		uint16_t count;
 	};
 
 	//void setup
 
-	std::vector<RLEbits> rleEncode(const std::vector<int>& data);
+	std::vector<RLEbits> rleEncode(const std::vector<uint8_t>& data);
 	//base station
 
-	std::vector<int> rleDecode(const std::vector<RLEbits>& code);
-	
+	std::vector<uint8_t> rleDecode(const std::vector<RLEbits>& code);
+
